limit frame ROI improvement when switching binning
This commit fixes an issue where the LimitFrame coordinates
could change when switching from bin 1 => bin 2 => bin 1 and
the initial bin 1 limit frame coordinates were not all even-valued.

The fix is to save all the per-binning limit frames in the profile
and load the corresponding value when switching binning (rather
than computing the value at the time of the binning change).